backend_django/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==4.1.1
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
gunicorn==20.1.0
jellyfish==0.9.0
numpy==1.23.3
pandas==1.4.4
pymongo==4.2.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.2
django-debug-toolbar==3.7.0
gtts==2.2.4

